added samples test project and renamed values controller
