+++ .github/workflows/upgrade-samples.yml
@@ -1,9 +1,8 @@
 name: CI
 
 on: 
-  pull_request:
-    branches: 
-      - develop
+  push:
+    branches:
       - master
 
 jobs:
@@ -15,5 +14,5 @@ jobs:
       uses: actions/setup-dotnet@v1
       with:
         dotnet-version: 3.1.100
-    - name: Run build.ps1
-      run: pwsh './build/build.ps1'
+    - name: Run upgrade-samples.ps1
+      run: pwsh './build/upgrade-samples.ps1'
