+++ BoltOn.sln
@@ -30,6 +30,8 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "BoltOn.Overrides", "samples
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "BoltOn.Templates", "src\BoltOn.Templates\BoltOn.Templates.csproj", "{25489C69-E070-4792-8F2D-E2870240B784}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "BoltOn.Samples.Tests", "samples\BoltOn.Samples.Tests\BoltOn.Samples.Tests.csproj", "{938B362E-1D48-43A2-A517-494A3331F9AE}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -80,6 +82,10 @@ Global
 		{25489C69-E070-4792-8F2D-E2870240B784}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{25489C69-E070-4792-8F2D-E2870240B784}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{25489C69-E070-4792-8F2D-E2870240B784}.Release|Any CPU.Build.0 = Release|Any CPU
+		{938B362E-1D48-43A2-A517-494A3331F9AE}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{938B362E-1D48-43A2-A517-494A3331F9AE}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{938B362E-1D48-43A2-A517-494A3331F9AE}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{938B362E-1D48-43A2-A517-494A3331F9AE}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -96,6 +102,7 @@ Global
 		{21DD4150-15A7-4417-9538-51F0B07FD31A} = {CF6C3199-B58A-4FB2-8E69-4E27B3A7BA29}
 		{7C28E882-C0E0-47C1-AD40-8EBDE16B5F43} = {15E5CCAC-85EB-4C98-B1FA-B5DEFD213FCD}
 		{25489C69-E070-4792-8F2D-E2870240B784} = {CF6C3199-B58A-4FB2-8E69-4E27B3A7BA29}
+		{938B362E-1D48-43A2-A517-494A3331F9AE} = {15E5CCAC-85EB-4C98-B1FA-B5DEFD213FCD}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {530F371A-8A38-41FE-9A1E-A03ACCDE867B}
